requirements.txt
playwright>=1.48
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
phonenumbers>=8.13
flask>=2.3
mysql-connector-python>=8.0